Enemies now hurt the player periodically instead of disappearing

+++ src/main.cpp
@@ -89,8 +89,8 @@ int main()
         PlayerElement* player = level1.getPlayerElement();
         if (player)
         {
-            g.driver->draw2DRectangle({255,255,0,0}, {5, 5, player->getHealth()*2, 15});
-            g.driver->draw2DRectangle({255,255,255,0}, {5, 20, player->getRewards()*20, 30});
+            g.driver->draw2DRectangle({255,255,0,0}, {5, 5, (s32)player->getHealth()*2, 15});
+            g.driver->draw2DRectangle({255,255,255,0}, {5, 20, (s32)player->getRewards()*20, 30});
         }
 
         g.driver->endScene();
